refactor: clarify span annotation summary naming
