+++ js/app/src/agent/tools/batchSpanAnnotate/applySpanAnnotations.ts
@@ -115,7 +115,7 @@ function commitCreateSpanAnnotations({
               span {
                 id
                 __typename
-                ...AnnotationSummaryGroup
+                ...SpanAnnotationSummaryGroup
                 ...SpanAnnotationsEditor_spanAnnotations
                   @arguments(filterUserIds: $filterUserIds)
                 ...SpanAnnotationsTable_annotations
